Safari-611.1.10/LayoutTests/media/restricted-audio-playback-with-document-gesture-expected.txt

Test that, when RequireUserGestureForAudioRateChange is set along with MediaUserGestureInheritsForDocument, a user gesture in the document will allow future media playback requests to succeed.


RUN(internals.setMediaElementRestrictions(mediaElement, "RequireUserGestureForAudioRateChange"))
RUN(internals.settings.setMediaUserGestureInheritsFromDocument(true))
Running with keyDown.
RUN(mediaElement.src = findMediaFile('audio', 'content/test'))
EVENT(canplaythrough)
RUN(mediaElement.play())
EXPECTED (mediaElement.paused == 'false') OK
END OF TEST
